requirements.txt
blinker==1.9.0
bottle==0.13.3
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
proxy_tools==0.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
pywebview==5.4
QtPy==2.4.3
typing_extensions==4.13.2
Werkzeug==3.1.3
